pyproject.toml
[project]
name = "stepsketch"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.10"
dependencies = [
    "google-genai>=1.59.0",
    "pydantic>=2.12.5",
    "pyside6>=6.10.1",
]

[project.optional-dependencies]
dev = [
    "nuitka>=2.3.0",
]

[project.scripts]
start = "src.main:main"
